Reduce number of allocations.
